Revert work on wand durability [https://trello.com/c/sOEzve54]
All wands now work once in survival, or infinitely in creative
