+++ src/main/java/org/sapphon/minecraft/modding/minecraftpython/spells/AbstractSpell.java
@@ -23,11 +23,6 @@ public abstract class AbstractSpell implements ISpell {
     }
 
     @Override
-    public int getMaximumUses() {
-        return getIntMetadataOrZero(SpellMetadataConstants.KEY_MAXIMUM_USES);
-    }
-
-    @Override
     public int getRequiredExperienceLevels() {
         return getIntMetadataOrZero(SpellMetadataConstants.KEY_REQUIRED_EXPERIENCE_LEVEL);
     }
